refactor(TableGroupService): 패키지 이동
