+++ src/main/java/kitchenpos/order/dto/OrderTableNumberOfGuestsRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ordertable.dto;
+package kitchenpos.order.dto;
 
 public class OrderTableNumberOfGuestsRequest {
     private int numberOfGuests;
