+++ src/main/java/kitchenpos/order/repository/OrderTableDao.java
@@ -1,10 +1,10 @@
-package kitchenpos.ordertable.repository;
+package kitchenpos.order.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.order.domain.OrderTable;
 
 public interface OrderTableDao extends JpaRepository<OrderTable, Long> {
     List<OrderTable> findAllByTableGroupId(Long tableGroupId);
