+++ src/main/java/kitchenpos/tablegroup/event/UnGroupedEvent.java
@@ -1,0 +1,15 @@
+package kitchenpos.tablegroup.event;
+
+import kitchenpos.tablegroup.domain.TableGroup;
+
+public class UnGroupedEvent {
+    private final TableGroup tableGroup;
+
+    public UnGroupedEvent(TableGroup tableGroup) {
+        this.tableGroup = tableGroup;
+    }
+
+    public TableGroup getTableGroup() {
+        return tableGroup;
+    }
+}
