+++ src/test/java/kitchenpos/order/domain/NumberOfGuestsTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ordertable.domain;
+package kitchenpos.order.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
