+++ src/test/java/kitchenpos/order/domain/OrderTableTest.java
@@ -1,6 +1,7 @@
-package kitchenpos.ordertable.domain;
+package kitchenpos.order.domain;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
